Local develop (#3)
* :memo: Add Gitter badge

* :memo: Add link to the Gitter in the ISSUE_TEMPLATE

* :bug: Fix loopFunction being called too early before MQTT connection is established - closes #260 (#290)

* :shirt: Fix lint

* :sparkles: Add destructor to HomieNode, that will abort() (#286)

* Added destructor to HomieNode, that will abort()

* :art: Log wifi client IP address upon connection. (#280)

This removes the need to hunt the device IP when debugging connectivity issues.

* :memo: Mention troubleshooting page in issue template

* :shirt: Fix cpplint failure, causing travis CI to be red (#291)

E.g. https://travis-ci.org/marvinroger/homie-esp8266/jobs/199213998

Ran the following command successfully:
  cpplint --repository=. --recursive \
  --filter=-whitespace/line_length,-legal/copyright,-runtime/printf,-build/include,-build/namespace,-runtime/int \
  ./src

* :fire: Remove MQTT ack logging

* :racehorse: Make tiny optimizations

* :art: Add a comma to the wifi connected log

* :art: Reorganize next boot mode feature

* :racehorse: Check length on input fields (#292)

* Fixing loopFunction being called too early before MQTT connection is established

* Started cleanup of strcpy/sprintf to include length check.

* Fixed warning under Interface.cpp

* :shirt: Fix lint

* :art: Print firmware name and version at boot

* :bug: Fix DeviceId incomplete MAC address (#296)

snprintf works with n-1 characters

expected behaviour: DeviceId is 12 characters
behaviour: only 11 characters are returned

(resulting in a "incomplete" mac address)

* :bug Fix millis() overflow in uptime - fix #299 (#302)

Corrected type declaration for correct overflow handling of millis
function in uptime calculation

+++ src/Homie/Uptime.hpp
@@ -11,6 +11,6 @@ class Uptime {
 
  private:
   uint64_t _seconds;
-  uint64_t _lastTick;
+  uint32_t _lastTick;
 };
 }  // namespace HomieInternals
